Move Solver.add_hints to VariablesManager
