+++ pyjobshop/ortools/VariablesManager.py
@@ -8,6 +8,7 @@ from ortools.sat.python.cp_model import (
 )
 
 from pyjobshop.ProblemData import ProblemData
+from pyjobshop.Solution import Solution
 from pyjobshop.utils import compute_min_max_durations
 
 
@@ -293,3 +294,44 @@ class VariablesManager:
             variables.append(SequenceVar(alt_vars))
 
         return variables
+
+    def add_hints(self, solution: Solution):
+        """
+        Adds hints to variables based on the given solution.
+        """
+        m, data = self._model, self._data
+        job_vars, task_vars, task_alt_vars = (
+            self.job_vars,
+            self.task_vars,
+            self.task_alt_vars,
+        )
+
+        m.clear_hints()
+
+        for idx in range(data.num_jobs):
+            job = data.jobs[idx]
+            job_var = job_vars[idx]
+            sol_tasks = [solution.tasks[task] for task in job.tasks]
+
+            job_start = min(task.start for task in sol_tasks)
+            job_end = max(task.end for task in sol_tasks)
+
+            m.add_hint(job_var.start, job_start)
+            m.add_hint(job_var.duration, job_end - job_start)
+            m.add_hint(job_var.end, job_end)
+
+        for idx in range(data.num_tasks):
+            task_var = task_vars[idx]
+            sol_task = solution.tasks[idx]
+
+            m.add_hint(task_var.start, sol_task.start)
+            m.add_hint(task_var.duration, sol_task.duration)
+            m.add_hint(task_var.end, sol_task.end)
+
+        for (task_idx, machine_idx), var in task_alt_vars.items():
+            sol_task = solution.tasks[task_idx]
+
+            m.add_hint(var.start, sol_task.start)
+            m.add_hint(var.duration, sol_task.duration)
+            m.add_hint(var.end, sol_task.end)
+            m.add_hint(var.is_present, machine_idx == sol_task.machine)
